==> Titanic/2.Feature Engineering.ipynb <==
import numpy as np
import pandas as pd
import string 
from patsy import dmatrices
from operator import itemgetter
#model 
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier, ExtraTreesClassifier, GradientBoostingRegressor 
from sklearn.linear_model import LogisticRegression, LinearRegression
from sklearn.svm import SVC
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split, StratifiedShuffleSplit
from sklearn.pipeline import Pipeline
from sklearn import preprocessing
import xgboost as xgb
#evaluation metrics 
from sklearn.metrics import classification_report
from sklearn.externals import joblib
#visulisation 
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline 

## 1. Read Files

train_df = pd.read_csv('data/train.csv')
test_df = pd.read_csv('data/test.csv')

train_df.columns

train_df.info()

test_df['Survived'] = 0
all_df = train_df.append(test_df)

test_df.shape

## 2. Feature Engineering - category features

Transform the category features 
1. Fare
2. Pclass
3. Family Size
4. Age group
5. Name length

def drop_col_not_req(df, cols):
    df.drop(cols, axis = 1, inplace = True)

def fare_category(fare):
    if (fare <= 4):
        return 'Very_Low_Fare'
    elif (fare <= 10):
        return 'Low_Fare'
    elif (fare <= 30):
        return 'Med_Fare'
    elif (fare <= 45):
        return 'High_Fare'
    else:
        return 'Very_High_Fare'

def pclass_fare_category(df, Pclass_1_mean_fare, Pclass_2_mean_fare, Pclass_3_mean_fare):
    if (df['Pclass'] == 1):
        if (df['Fare'] <= Pclass_1_mean_fare):
            return 'Pclass_1_Low_Fare'
        else:
            return 'Pclass_1_High_Fare'
    elif (df['Pclass'] == 2):
        if (df['Fare'] <= Pclass_2_mean_fare):
            return 'Pclass_2_Low_Fare'
        else:
            return 'Pclass_2_High_Fare'
    elif (df['Pclass'] == 3):
        if (df['Fare'] <= Pclass_3_mean_fare):
            return 'Pclass_3_Low_Fare'
        else:
            return 'Pclass_3_High_Fare'

def family_size_category(family_size):
    if (family_size <= 1):
        return 'Single'
    elif (family_size <= 3):
        return 'Small_Family'
    else:
        return 'Large_Family'

def age_group_cat(age):
    if (age <= 1):
        return 'Baby'
    if (age <= 4):
        return 'Toddler'
    elif(age <= 12):
        return 'Child'
    elif (age <= 19):
        return 'Teenager'
    elif (age <= 30):
        return 'Adult'
    elif (age <= 50):
        return 'Middle_Aged'
    elif(age < 60):
        return 'Senior_Citizen'
    else:
        return 'Old'

def name_len_category(name_len):
    if (name_len <= 19):
        return 'Very_Short_Name'
    elif (name_len <= 28):
        return 'Short_Name'
    elif (name_len <= 45):
        return 'Medium_Name'
    else:
        return 'Long_Name'

## 3. Handling missing values

Use **GradientBoostingRegressor** and **LinearRegression** to fill the missing value 

def fill_missing_age(missing_age_train, missing_age_test):
    missing_age_X_train = missing_age_train.drop(['Age'], axis = 1)
    missing_age_y_train = missing_age_train['Age']
    missing_age_X_test = missing_age_test.drop(['Age'], axis = 1)
    
    #gridsearch for best parameters fit for GradientBoostingRegressor
    gbm_reg = GradientBoostingRegressor(random_state = 42)
    gbm_reg_param_grid = {'n_estimators': [2000], 'max_depth': [3], 'learning_rate': [0.01], 'max_features': [3]}
    gbm_reg_grid = GridSearchCV(gbm_reg, gbm_reg_param_grid, cv = 10, n_jobs = 25, verbose = 1, scoring = 'neg_mean_squared_error')
    gbm_reg_grid.fit(missing_age_X_train, missing_age_y_train)
    
    print("Age feature Best GB Params: " + str(gbm_reg_grid.best_params_))
    print("Age feature Best GB Score: " + str(gbm_reg_grid.best_score_))
    print("GB Train Error for 'Age' Feature Regressor: " + str(gbm_reg_grid.score(missing_age_X_train, missing_age_y_train)))
    
    missing_age_test['Age_GB'] = gbm_reg_grid.predict(missing_age_X_test)
    print(missing_age_test['Age_GB'][:4])
    
    #gridsearch for best parameters fit for LinearRegression
    lrf_reg = LinearRegression()
    lrf_reg_param_grid = {'fit_intercept': [True], 'normalize': [True]}
    lrf_reg_grid = GridSearchCV(lrf_reg, lrf_reg_param_grid, cv = 10, n_jobs = 25, verbose = 1, scoring = 'neg_mean_squared_error')
    lrf_reg_grid.fit(missing_age_X_train, missing_age_y_train)
    
    print("Age feature Best LR Params: " + str(lrf_reg_grid.best_params_))
    print("Age feature Best LR Score: " + str(lrf_reg_grid.best_score_))
    print("LR Train Error for 'Age' Feature Regressor: " + str(lrf_reg_grid.score(missing_age_X_train, missing_age_y_train)))
    
    missing_age_test['Age_LRF'] = lrf_reg_grid.predict(missing_age_X_test)
    print(missing_age_test['Age_LRF'][:4])
    
    missing_age_test['Age'] = missing_age_test[['Age_GB', 'Age_LRF']].mean(axis = 1)

    print(missing_age_test['Age'][:4])
    drop_col_not_req(missing_age_test, ['Age_GB', 'Age_LRF'])

    return missing_age_test

## 4. Pick top 'N' features 

Pick up top 'N' features in different ensemble models (**RandomForestClassifier**, **AdaBoostClassifier**, and **ExtraTreesClassifier**)

def get_top_n_features(titanic_train_data_X, titanic_train_data_y, top_n_features):
    #random forest 
    rf_est = RandomForestClassifier(random_state = 42)
    rf_param_grid = {'n_estimators' : [500], 'min_samples_split':[2, 3], 'max_depth':[20]}
    rf_grid =GridSearchCV(rf_est, rf_param_grid, n_jobs = 25, cv = 10, verbose = 1)
    rf_grid.fit(titanic_train_data_X, titanic_train_data_y)
    
    print("Top N Features Best RF Params: " + str(rf_grid.best_params_))
    print("Top N Features Best RF Score: " + str(rf_grid.best_score_))
    print("Top N Features RF Train Error: " + str(rf_grid.score(titanic_train_data_X, titanic_train_data_y)))

    feature_imp_sorted_rf = pd.DataFrame({'feature': list(titanic_train_data_X), 
                                          'importance': rf_grid.best_estimator_.feature_importances_}).sort_values('importance', ascending = False)
    features_top_n_rf = feature_imp_sorted_rf.head(top_n_features)['feature']
    print("Sample 25 Features from RF Classifier:")
    print(str(features_top_n_rf[:25]))
    
    #ada boost 
    ada_est = AdaBoostClassifier(random_state = 42)
    ada_param_grid = {'n_estimators' : [500], 'learning_rate': [0.5, 0.6]}
    ada_grid = GridSearchCV(ada_est, ada_param_grid, n_jobs = 25, cv = 10, verbose = 1)
    ada_grid.fit(titanic_train_data_X, titanic_train_data_y)
    
    print("Top N Features Best Ada Params: " + str(ada_grid.best_params_))
    print("Top N Features Best Ada Score: " + str(ada_grid.best_score_))
    print("Top N Features Ada Train Error: " + str(ada_grid.score(titanic_train_data_X, titanic_train_data_y)))
    
    feature_imp_sorted_ada = pd.DataFrame({'feature': list(titanic_train_data_X), 'importance': ada_grid.best_estimator_.feature_importances_}).sort_values('importance', ascending = False)
    features_top_n_ada = feature_imp_sorted_ada.head(top_n_features)['feature']
    print("Sample 25 Features from Ada Classifier:")
    print(str(features_top_n_ada[:25]))
    
    #extra tree 
    et_est = ExtraTreesClassifier(random_state = 42)
    et_param_grid = {'n_estimators' : [500], 'min_samples_split':[3, 4], 'max_depth':[15]}
    et_grid = GridSearchCV(et_est, et_param_grid, n_jobs = 25, cv = 10, verbose = 1)
    et_grid.fit(titanic_train_data_X, titanic_train_data_y)
    
    print("Top N Features Best ET Params: " + str(et_grid.best_params_))
    print("Top N Features Best ET Score: " + str(et_grid.best_score_))
    print("Top N Features ET Train Error: " + str(et_grid.score(titanic_train_data_X, titanic_train_data_y)))
    
    feature_imp_sorted_et = pd.DataFrame({'feature': list(titanic_train_data_X), 'importance': et_grid.best_estimator_.feature_importances_}).sort_values('importance', ascending = False)
    features_top_n_et = feature_imp_sorted_et.head(top_n_features)['feature']
    print("Sample 25 Features from ET Classifier:")
    print(str(features_top_n_et[:25]))
    
    #### Merge top_n_features from all three models
    features_top_n = pd.concat([features_top_n_rf, features_top_n_ada, features_top_n_et], ignore_index = True).drop_duplicates()
    
    return features_top_n

## 5. Creating Features

### 5.1 Embarked

print(all_df.groupby(['Survived', 'Embarked'])['Survived'].count())
print(all_df['PassengerId'].groupby(by = all_df['Embarked']).count().sort_values(ascending = False))
print(all_df['Fare'].groupby(by = all_df['Embarked']).mean().sort_values(ascending = False))


if (all_df['Embarked'].isnull().sum() != 0):
    all_df['Embarked'].fillna(all_df['Embarked'].mode().iloc[0], inplace=True)

emb_dummies_df = pd.get_dummies(all_df['Embarked'],
                                prefix = all_df[['Embarked']].columns[0])
all_df = pd.concat([all_df, emb_dummies_df], axis = 1)
all_df.head(5)

### 5.2 Sex 

sex_dummies_df = pd.get_dummies(all_df['Sex'],
                                prefix = all_df[['Sex']].columns[0])
all_df = pd.concat([all_df, sex_dummies_df], axis = 1)
all_df.head(3)

### 5.3 Name Title

all_df['Title'] = all_df['Name'].str.extract('.+,(.+)').str.extract('^(.+?)\.').str.strip()
print(all_df['Title'].unique())
print(all_df['Title'].groupby(by = all_df['Title']).count().sort_values(ascending = False))

title_Dict = {}
title_Dict.update(dict.fromkeys(["Capt", "Col", "Major", "Dr", "Rev"], "Officer"))
title_Dict.update(dict.fromkeys(["Jonkheer", "Don", "Sir", "the Countess", "Dona", "Lady"], "Royalty"))
title_Dict.update(dict.fromkeys(["Mme", "Ms", "Mrs"], "Mrs"))
title_Dict.update(dict.fromkeys(["Mlle", "Miss"], "Miss"))
title_Dict.update(dict.fromkeys(["Mr", "Ms"], "Mr"))
title_Dict.update(dict.fromkeys(["Master"], "Master"))

all_df['Title'] = all_df['Title'].map(title_Dict)
print(all_df['Title'].groupby(by = all_df['Title']).count().sort_values(ascending = False))

title_dummies_df = pd.get_dummies(all_df['Title'],
                                prefix = all_df[['Title']].columns[0])
all_df = pd.concat([all_df, title_dummies_df], axis = 1)

### 5.4 Name length

all_df['Name_Length'] = all_df['Name'].str.len()
print(all_df['Name_Length'].groupby(by = all_df['Name_Length']).count().sort_values(ascending = False)[:5])

all_df['Name_Length_Category'] = all_df['Name_Length'].map(name_len_category)
print(all_df['Name_Length_Category'].groupby(by = all_df['Name_Length_Category']).count().sort_values(ascending = False))

le_fare = preprocessing.LabelEncoder()
le_fare.fit(np.array(['Very_Short_Name', 'Short_Name', 'Medium_Name', 'Long_Name', 'Very_High_Fare']))
all_df['Name_Length_Category'] = le_fare.transform(all_df['Name_Length_Category'])

print(all_df[['Name_Length_Category', 'Survived']].corr())

first_name_dummies_df = pd.get_dummies(all_df['Name_Length_Category'],
                                prefix = all_df[['Name_Length_Category']].columns[0])
all_df = pd.concat([all_df, first_name_dummies_df], axis = 1)
all_df.head(3)

### 5.5 Fare

#filling the missing value using mean in Fare
if (all_df['Fare'].isnull().sum() != 0):
    all_df['Fare'] = all_df[['Fare']].fillna(all_df.groupby('Pclass').mean(axis=1))

#average fare for the tickets 
all_df['Group_Ticket'] = all_df['Fare'].groupby(by = all_df['Ticket']).transform('count')
all_df['Fare'] = all_df['Fare']/all_df['Group_Ticket']
all_df.drop(['Group_Ticket'], axis = 1, inplace = True)
all_df.head(5)

#check if there are wired value, which is fare =0 
if (sum(n == 0 for n in all_df.Fare.values.flatten()) > 0):
    all_df.loc[all_df.Fare == 0, 'Fare'] = np.nan
    all_df['Fare'] = all_df[['Fare']].fillna(all_df.groupby('Pclass').transform('mean'))

all_df['Fare'].describe()

#transform fare to fare_category 
all_df['Fare_Category'] = all_df['Fare'].map(fare_category)
le_fare = preprocessing.LabelEncoder()
le_fare.fit(np.array(['Very_Low_Fare', 'Low_Fare', 'Med_Fare', 'High_Fare', 'Very_High_Fare']))
all_df['Fare_Category'] = le_fare.transform(all_df['Fare_Category'])

fare_cat_dummies_df = pd.get_dummies(all_df['Fare_Category'],
                                prefix = all_df[['Fare_Category']].columns[0])
all_df = pd.concat([all_df, fare_cat_dummies_df], axis = 1)

print(all_df['Fare_Category'].groupby(by = all_df['Fare_Category']).count().sort_values(ascending = False))

### 5.6 Pclass

print(all_df['Fare'].groupby(by = all_df['Pclass']).mean())
Pclass_1_mean_fare = all_df['Fare'].groupby(by = all_df['Pclass']).mean().get([1]).values[0]
Pclass_2_mean_fare = all_df['Fare'].groupby(by = all_df['Pclass']).mean().get([2]).values[0]
Pclass_3_mean_fare = all_df['Fare'].groupby(by = all_df['Pclass']).mean().get([3]).values[0]

#category variable from Pclass and Fare 
all_df['Pclass_Fare_Category'] = all_df.apply(pclass_fare_category, args=(Pclass_1_mean_fare, Pclass_2_mean_fare, Pclass_3_mean_fare), axis = 1)
print(all_df['Pclass_Fare_Category'].groupby(by = all_df['Pclass_Fare_Category']).count().sort_values(ascending = False))

le_fare = preprocessing.LabelEncoder()
le_fare.fit(np.array(['Pclass_1_Low_Fare', 'Pclass_1_High_Fare', 'Pclass_2_Low_Fare', 'Pclass_2_High_Fare', 'Pclass_3_Low_Fare', 'Pclass_3_High_Fare']))
all_df['Pclass_Fare_Category'] = le_fare.transform(all_df['Pclass_Fare_Category'])

print(all_df['Fare'].groupby(by = all_df['Pclass']).mean().sort_values(ascending = True))
all_df['Pclass'].replace([1, 2, 3],[Pclass_1_mean_fare, Pclass_2_mean_fare, Pclass_3_mean_fare], inplace = True)


### 5.7 Parch and SibSp

all_df['Family_Size'] = all_df['Parch'] + all_df['SibSp'] + 1
print(all_df['Family_Size'].groupby(by = all_df['Family_Size']).count().sort_values(ascending = False))

all_df['Family_Size_Category'] = all_df['Family_Size'].map(family_size_category)

print(all_df['Family_Size_Category'].groupby(by = all_df['Family_Size_Category']).count().sort_values(ascending = False))
print(all_df.groupby(['Survived', 'Family_Size_Category'])['Survived'].count())

le_family = preprocessing.LabelEncoder()
le_family.fit(np.array(['Single', 'Small_Family', 'Large_Family']))
all_df['Family_Size_Category'] = le_family.transform(all_df['Family_Size_Category'])

fam_size_cat_dummies_df = pd.get_dummies(all_df['Family_Size_Category'],
                                prefix = all_df[['Family_Size_Category']].columns[0])
all_df = pd.concat([all_df, fam_size_cat_dummies_df], axis = 1)

### 5.8 Age

print(all_df['Age'].groupby(by = all_df['Title']).mean().sort_values(ascending = True))

#flag the age null to 1 and not null in 0 
all_df['Age_Null'] = all_df['Age'].apply(lambda x: 1 if(pd.notnull(x)) else 0)

#prepare the dataframe for training 
missing_age_df = pd.DataFrame(all_df[['Age', 'Parch', 'Sex', 'SibSp', 'Family_Size', 'Family_Size_Category', 'Title', 'Fare', 'Fare_Category', 'Pclass', 'Embarked']])
missing_age_df = pd.get_dummies(missing_age_df, columns = ['Title', 'Family_Size_Category', 'Fare_Category', 'Sex', 'Pclass', 'Embarked'])
missing_age_df.shape
missing_age_df.info()

missing_age_train = missing_age_df[missing_age_df['Age'].notnull()]
missing_age_test  = missing_age_df[missing_age_df['Age'].isnull()]


#use the before function to fill the missing Age 
all_df.loc[(all_df.Age.isnull()), 'Age']= fill_missing_age(missing_age_train, missing_age_test)


print(all_df['Age'].groupby(by = all_df['Title']).mean().sort_values(ascending = True))

all_df['Age_Category'] = all_df['Age'].map(age_group_cat)
le_age = preprocessing.LabelEncoder()
le_age.fit(np.array(['Baby', 'Toddler', 'Child', 'Teenager', 'Adult', 'Middle_Aged', 'Senior_Citizen', 'Old']))
all_df['Age_Category'] = le_age.transform(all_df['Age_Category'])

age_cat_dummies_df = pd.get_dummies(all_df['Age_Category'],
                                prefix = all_df[['Age_Category']].columns[0])
all_df = pd.concat([all_df, age_cat_dummies_df], axis = 1)

### 5.9 Ticket & Cabin

all_df['Ticket_Letter'] = all_df['Ticket'].str.split().str[0]
all_df['Ticket_Letter'] = all_df['Ticket_Letter'].apply(lambda x: np.NaN if x.isnumeric() else x)
all_df['Ticket_Number'] = all_df['Ticket'].apply(lambda x: pd.to_numeric(x, errors='coerce'))
all_df['Ticket_Number'].fillna(0, inplace = True)
all_df = pd.get_dummies(all_df, columns = ['Ticket', 'Ticket_Letter'])
all_df.shape

all_df['Cabin_Letter'] = all_df['Cabin'].apply(lambda x: str(x)[0]  if(pd.notnull(x)) else x)
all_df = pd.get_dummies(all_df, columns = ['Cabin', 'Cabin_Letter'])
all_df.shape

### 5.10 Normalize Age and Fare

all_df[['Age', 'Fare']][:5]

scale_age_fare = preprocessing.StandardScaler().fit(all_df[['Age', 'Fare']])
all_df[['Age', 'Fare']] = scale_age_fare.transform(all_df[['Age', 'Fare']])

all_df[['Age', 'Fare']].mean()

### 5.11 Drop columns 

all_df.drop(['Name', 'PassengerId', 'Embarked', 'Sex', 'Title', 'Fare_Category', 'Family_Size_Category',
               'Age_Category', 'Name_Length_Category'], 
              axis = 1, inplace = True)

train_data = all_df[:891]
test_data = all_df[891:]

titanic_train_data_X = train_data.drop(['Survived'], axis = 1)
titanic_train_data_y = train_data['Survived']

titanic_test_data_X = test_data.drop(['Survived'], axis = 1)

titanic_test_data_X.shape

## 6. Choose Top features

features_to_pick = 200
features_top_n = get_top_n_features(titanic_train_data_X, titanic_train_data_y, features_to_pick)

print("Total Features: " + str(all_df.shape))
print("Picked Features: " + str(features_top_n.shape))

titanic_train_data_X = titanic_train_data_X[features_top_n]
print(titanic_train_data_X.shape)
print(titanic_train_data_X.info())

titanic_test_data_X = titanic_test_data_X[features_top_n]
titanic_test_data_X.shape
titanic_test_data_X.info()

titanic_train_data = pd.concat([titanic_train_data_X, titanic_train_data_y], axis = 1)

titanic_train_data_X.shape

titanic_train_data.shape

titanic_test_data_X.to_csv("data/topfeature_test.csv", index=False)
titanic_train_data.to_csv("data/topfeature_train.csv", index=False)
