==> Titanic/4.Model-Voting.ipynb <==
import pandas as pd 
import numpy as np 
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier, ExtraTreesClassifier
from sklearn.ensemble import VotingClassifier, AdaBoostClassifier
import xgboost as xgb
from sklearn.svm import SVC 


ls

df_all = pd.read_csv('data/test.csv')
topfeature_train = pd.read_csv('data/topfeature_train.csv')
topfeature_test_X = pd.read_csv('data/topfeature_test.csv')

topfeature_train_X = topfeature_train.drop(['Survived'], axis = 1)
topfeature_train_y = topfeature_train['Survived']

print(df_all.shape)

print(topfeature_train_X.shape)
print(topfeature_train_y.shape)
print(topfeature_test_X.shape)

rf_est = RandomForestClassifier(n_estimators = 750, criterion = 'gini', max_features = 'sqrt',
                                max_depth = 3, min_samples_split = 4, min_samples_leaf = 2, n_jobs = 50,
                                random_state = 42, verbose = 1)

gbm_est = GradientBoostingClassifier(n_estimators = 1000, learning_rate = 0.0008,
                                     loss = 'exponential', min_samples_split = 3,
                                     min_samples_leaf = 2, max_features ='sqrt', max_depth = 3,
                                     random_state = 42, verbose = 1)

et_est = ExtraTreesClassifier(n_estimators = 750, max_features = 'sqrt', max_depth = 35,
                              n_jobs = 50, criterion = 'entropy', random_state = 42, verbose = 1)

#gbm = xgb.XGBClassifier(max_depth=3, n_estimators=300, learning_rate=0.05)

voting_est = VotingClassifier(estimators = [('rf', rf_est),('gbm', gbm_est),('et', et_est)],
                              voting = 'soft', weights = [3,5,2],n_jobs = 50)
                                       
voting_est.fit(topfeature_train_X, topfeature_train_y)
print("VotingClassifier Score: " + str(voting_est.score(topfeature_train_X, topfeature_train_y)))
print("VotingClassifier Estimators: " + str(voting_est.estimators_))

topfeature_test_X['Survived'] = voting_est.predict(topfeature_test_X)

submission = pd.DataFrame({'PassengerId': df_all.loc[:, 'PassengerId'],
                           'Survived': topfeature_test_X.loc[:, 'Survived'].astype(np.int32)})
submission.to_csv("submission/submission08_08_1.csv", index = False)

submission.isnull().any()

df_tset =
