==> Titanic/5.Model-stacking.ipynb <==
import pandas as pd 
import numpy as np 
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier, ExtraTreesClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.svm import SVC 
import xgboost as xgb
from sklearn.model_selection import StratifiedShuffleSplit, GridSearchCV


df_all = pd.read_csv('data/test.csv')
topfeature_train = pd.read_csv('data/topfeature_train.csv')
topfeature_test_X = pd.read_csv('data/topfeature_test.csv')

topfeature_train_X = topfeature_train.drop(['Survived'], axis = 1)
topfeature_train_y = topfeature_train['Survived'].ravel()

## Base level model 

#Random Forest
rf_params ={
    'n_estimators':[500], 
    'criterion':['gini'], 
    'max_features': ['sqrt'],
    'max_depth':[6,20], 
    'min_samples_split':[3], 
    'min_samples_leaf':[2], 
    'n_jobs':[50], 
    'random_state':[42], 
    'verbose':[0]
}
#rf_param_grid = {'n_estimators' : [500], 'min_samples_split':[2], 'max_depth':[20]}
rf_clf = RandomForestClassifier()

#Gradient Boosting
gbm_params={
    'n_estimators': [500], 
    'learning_rate': [0.0008,0.75],
    'loss': ['exponential'], 
    'min_samples_split': [3],
    'min_samples_leaf': [2], 
    'max_features': ['sqrt'], 
    'max_depth': [3],
    'random_state': [42], 
    'verbose': [0,1]
}
gbm_est = GradientBoostingClassifier()

#Extra Trees
et_est = ExtraTreesClassifier(n_estimators = 750, max_features = 'sqrt', max_depth = 35,
                              n_jobs = 50, criterion = 'entropy', random_state = 42, verbose = 1)

#Ada Boosting 

#SVC

**Stratified shuffle split** + **Gridsearch** 

stratifiedCV = StratifiedShuffleSplit(n_splits = 10, test_size=0.2, random_state =0)

def grid_cv(clf,param,name):
    grid_search = GridSearchCV(clf,param,verbose = 3, scoring ='accuracy', cv = stratifiedCV)
    grid_search.fit(topfeature_train_X,topfeature_train_y)
    print(name, "Best Params:" + str(grid_search.best_params_))
    print(name, "Best Score:" + str(grid_search.best_score_))
    name = grid_search.predict(topfeature_test_X)

    
grid_cv(rf_clf, rf_params, 'randomForest')

randomForest
